+++ src/app/beer_garden/router.py
@@ -294,12 +294,34 @@ def execute_local(operation: Operation):
 
 def update_api_heartbeat(operation: Operation):
     if (
-        operation.source_garden_name is not None
-        and operation.source_garden_name != config.get("garden.name")
-        and operation.source_api is not None
+        operation.source_api is not None
         and operation.operation_type == "PUBLISH_EVENT"
         and operation.model.name == Events.GARDEN_SYNC.name
     ):
+        if operation.source_garden_name == config.get("garden.name"):
+
+            if operation.model.payload.name != operation.source_garden_name:
+
+                local_garden = get_garden(config.get("garden.name"))
+
+                # Will only support mapping 1 hop away legacy Garden Syncs
+                child_garden = False
+                for child in local_garden.children:
+                    if child.name == operation.model.payload.name:
+                        logger.warning(
+                            (
+                                "Legacy (3.16 or prior) GARDEN_SYNC operation "
+                                f"seen for Beer-Garden '{operation.model.payload.name}'"
+                            )
+                        )
+                        operation.source_garden_name = operation.model.payload.name
+                        child_garden = True
+                        break
+                if child_garden:
+                    return
+            else:
+                return
+
         source_garden = getattr(gardens, operation.source_garden_name, None)
         if operation.model.payload.name == operation.source_garden_name:
             beer_garden.garden.check_garden_receiving_heartbeat(