+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TBD
 - Added support to prune legacy request with no command type
 - Added support for group/groups in beer.conf and group environment variable
 - Added max_concurrent support for config.yml under `plugin.local.max_concurrent`
+- Added backwards compatibility for Garden Syncs for 3.16 or older releases
 - Added max concurrent support for beer.conf under `MAX_CONCURRENT`
 - Added feature to cancel an ongoing request or delete a completed request
 - Fixed API swagger docs for Job execution